LayoutTests/platform/chromium-mac/fast/dom/xmlhttprequest-constructor-in-detached-document-expected.txt

CONSOLE MESSAGE: line 8: Uncaught TypeError: undefined is not a function
Text for bug 25290: Crash when constructing XMLHttpRequest in a detached document.

PASS
